requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

streamlit>=1.30.0
requests>=2.31.0
azure-functions
openai
pydantic
pydantic-settings